+++ presentation/src/main/java/com/agileburo/anytype/presentation/desktop/DashboardView.kt
@@ -1,6 +1,12 @@
 package com.agileburo.anytype.presentation.desktop
 
+import com.agileburo.anytype.domain.common.Id
+
 sealed class DashboardView {
-    data class Document(val id: String, val title: String?, val emoji: String? = null) :
-        DashboardView()
+    data class Document(
+        val id: Id,
+        val target: Id,
+        val title: String? = null,
+        val emoji: String? = null
+    ) : DashboardView()
 }